+++ wct/layers/conv.py
@@ -1,0 +1,13 @@
+import tensorflow as tf
+from keras.layers import Conv2D, Lambda
+
+def reflect(x, pad=1):
+    return tf.pad(x, [[0, 0], [pad, pad], [pad, pad], [0, 0]], 'REFLECT')
+
+def ReflectingConv2D(*args, **kwargs):
+    def reflect_convolve(x):
+        x = reflect(x)
+        x = Conv2D(*args, **kwargs)(x)
+        return x
+        
+    return Lambda(reflect_convolve, name=kwargs['name'])
